+++ torch_brain/dataset/__init__.py
@@ -1,13 +1,3 @@
-from .dataset import (
-    Dataset,
-    NestedDataset,
-    SpikingDatasetMixin,
-    NestedSpikingDataset,
-    DatasetIndex,
-)
-
-from .PerichMillerPopulation2018 import PerichMillerPopulation2018
-from .PeiPandarinathNLB2021 import PeiPandarinathNLB2021
-from .FlintSlutzkyAccurate2012 import FlintSlutzkyAccurate2012
-from .ChurchlandShenoyNeural2012 import ChurchlandShenoyNeural2012
-from .OdohertySabesNonhuman2017 import OdohertySabesNonhuman2017
+from .dataset import Dataset, DatasetIndex
+from .nested import NestedDataset, NestedSpikingDataset
+from .mixins import SpikingDatasetMixin
