+++ torch_brain/dataset/mixins.py
@@ -1,0 +1,7 @@
+import numpy as np
+
+
+class SpikingDatasetMixin:
+    def get_unit_ids(self) -> list[str]:
+        ans = [self.get_recording(rid).units.id for rid in self.recording_ids]
+        return np.sort(np.concatenate(ans)).tolist()
