+++ torch_brain/datasets/__init__.py
@@ -1,0 +1,5 @@
+from .PerichMillerPopulation2018 import PerichMillerPopulation2018
+from .PeiPandarinathNLB2021 import PeiPandarinathNLB2021
+from .FlintSlutzkyAccurate2012 import FlintSlutzkyAccurate2012
+from .ChurchlandShenoyNeural2012 import ChurchlandShenoyNeural2012
+from .OdohertySabesNonhuman2017 import OdohertySabesNonhuman2017
